+++ src/Simd/SimdSynetConvolution32f.h
@@ -142,9 +142,9 @@ namespace Simd
             return ss.str();
         }
 
-        SIMD_INLINE long long Flop() const
+        SIMD_INLINE int64_t Flop() const
         {
-            return batch* kernelY* kernelX* srcC* dstH* dstW* dstC / group * 2;
+            return int64_t(batch) * kernelY * kernelX * srcC * dstH * dstW * dstC / group * 2;
         }
 #endif
     };
